moved tls config from transport.go to utils.go
