+++ util.go
@@ -1,6 +1,7 @@
 package sentry
 
 import (
+	"crypto/tls"
 	"encoding/json"
 	"fmt"
 	"os"
@@ -108,4 +109,25 @@ func revisionFromBuildInfo(info *debug.BuildInfo) string {
 
 func Pointer[T any](v T) *T {
 	return &v
+}
+
+func getTLSConfig(options ClientOptions) *tls.Config {
+	if options.TLSConfig == nil && options.CaCerts == nil {
+		return nil
+	}
+
+	var tlsConfig *tls.Config
+	if options.TLSConfig != nil {
+		tlsConfig = options.TLSConfig.Clone()
+	} else {
+		// #nosec G402 -- We should be using `MinVersion: tls.VersionTLS12`,
+		// 				 but we don't want to break peoples code without the major bump.
+		tlsConfig = &tls.Config{}
+	}
+
+	if tlsConfig.RootCAs == nil && options.CaCerts != nil {
+		tlsConfig.RootCAs = options.CaCerts
+	}
+
+	return tlsConfig
 }
